BoardUtils - add print function to print a long as 8x8 as board of its bits. FenTranslator - changed value to byte. SpecialMovesHandler - new class to handle everything regarding special moves - en passant and castling. updateSpecialMoves will update status of the special moves after a move. getSpecialMoves - will add the moves, now demo. TestingMain - removed. ChessGame - class that represent a chess game with all of its attributes, responsible for executing moves, checking for checks/checkmate and giving the legals moves a piece can do give a square. removeIllegalMoves still demo. tempGui - changed to use ChessGame

+++ src/FenTranslator.java
@@ -155,7 +155,7 @@ public class FenTranslator {
         return blackLongCastle;
     }
 
-    public long getEnPassantSquareToCapture() {
+    public byte getEnPassantSquareToCapture() {
         return enPassantTargetSquare;
     }
